test: move tests to fuzz subdirectory (noop)
