+++ test/extensions/filters/listener/common/fuzz/listener_filter_fuzzer.cc
@@ -1,4 +1,4 @@
-#include "test/extensions/filters/listener/common/listener_filter_fuzzer.h"
+#include "test/extensions/filters/listener/common/fuzz/listener_filter_fuzzer.h"
 
 #include "common/network/utility.h"
