+++ test/extensions/filters/listener/original_dst/BUILD
@@ -28,6 +28,6 @@ envoy_cc_fuzz_test(
     corpus = "original_dst_corpus",
     deps = [
         "//source/extensions/filters/listener/original_dst:original_dst_lib",
-        "//test/extensions/filters/listener/common:listener_filter_fuzzer_lib",
+        "//test/extensions/filters/listener/common/fuzz:listener_filter_fuzzer_lib",
     ],
 )
