+++ test/extensions/filters/listener/original_src/original_src_fuzz_test.proto
@@ -3,7 +3,7 @@ syntax = "proto3";
 package envoy.extensions.filters.listener.original_src;
 
 import "envoy/extensions/filters/listener/original_src/v3/original_src.proto";
-import "test/extensions/filters/listener/common/listener_filter_fuzzer.proto";
+import "test/extensions/filters/listener/common/fuzz/listener_filter_fuzzer.proto";
 import "validate/validate.proto";
 
 message OriginalSrcTestCase {
